extract exit_button
